add hook for swr cache mutation
