+++ src/mutations/use-set-notification-state-mutation.ts
@@ -1,7 +1,7 @@
 import { gql } from 'graphql-request'
-import { useSWRConfig, mutate } from 'swr'
 
 import { useMutationFetch } from './helper/use-mutation-fetch'
+import { useSWRCacheMutate } from './helper/use-swr-cache-mutate'
 import { NotificationSetStateInput } from '@/fetcher/graphql-types/operations'
 
 const mutation = gql`
@@ -16,38 +16,22 @@ const mutation = gql`
 
 export function useSetNotificationStateMutation() {
   const mutationFetch = useMutationFetch()
-  const { cache } = useSWRConfig()
+  const mutateSWRCache = useSWRCacheMutate()
 
   return async function (input: NotificationSetStateInput) {
     const success = await mutationFetch(mutation, input)
 
-    // note: Maybe implement global cache key management, but this works okay
+    // not very elegant and cache names could change in the future,
+    // but for now it works quite well
 
-    if (success) {
-      if (!(cache instanceof Map)) {
-        throw new Error(
-          'matchMutate requires the cache provider to be a Map instance'
-        )
-      }
-
-      const keys = []
-
-      for (const key of cache.keys() as IterableIterator<string>) {
-        const shouldBeMutated =
-          (key.startsWith('arg@') &&
-            key.indexOf('query notifications($first') > 0 &&
-            key.indexOf('"unread":true') > 0) ||
-          key.indexOf('notifications(unread: true)') > 0
-
-        if (shouldBeMutated) {
-          keys.push(key)
-        }
-      }
-
-      keys.forEach((key) => {
-        void mutate(key)
-      })
-    }
+    mutateSWRCache((key: string) => {
+      return (
+        (key.startsWith('arg@') &&
+          key.indexOf('query notifications($first') > 0 &&
+          key.indexOf('"unread":true') > 0) ||
+        key.indexOf('notifications(unread: true)') > 0
+      )
+    })
     return success
   }
 }
